Fixat så update metoden för observers så att utskrifter skrivs då vi notifierar observer
Har även lagt till korrekt formatering vid utskrifterna med hjälp av metoden customFormat

Nu ger utskrifterna precis samma som på instruktionen!

+++ src/state/SupermarketState.java
@@ -145,7 +145,7 @@ public class SupermarketState extends SimState {
 	 * Uppdaterar den tiden som kassorna varit lediga
 	 */
 	public void updateFreeCashierTime() {
-		this.sumTimeFreeCheckouts += (double) getFreeCashiers() * (getTime()-getPreviousTime());
+		this.sumTimeFreeCheckouts += getFreeCashiers() * (getTime()-getPreviousTime());
 	}
 	
 	/**
